% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExprSimulato.R
\name{Add.Cell.Cell.Int}
\alias{Add.Cell.Cell.Int}
\title{Add.Cell.Cell.Int}
\usage{
Add.Cell.Cell.Int(
  ppp.obj,
  sim.count,
  dist.threshold = 0.05,
  Gaussian.effect.size.parameter = c(2, 0.1),
  interacting.cell.types = NULL,
  mech1.prop = 0.1,
  mech2.prop = 0.2,
  mech3.prop = 0.1,
  seed = NULL,
  bond.extreme = T
)
}
\arguments{
\item{ppp.obj}{PointLoc}

\item{sim.count}{PointLoc}

\item{dist.threshold}{PointLoc}

\item{Gaussian.effect.size.parameter}{PointLoc}

\item{interacting.cell.types}{interacting.cell.types=list(c("ExcitatoryNeuron", "Astrocyte"));interacting.cell.types=list(c("ExcitatoryNeuron", "Astrocyte"), c("ExcitatoryNeuron", "Oligodendrocyte"))}

\item{mech1.prop}{PointLoc}

\item{mech2.prop}{PointLoc}

\item{mech3.prop}{PointLoc}
}
\value{
\item{log.sim.change1:}{ppp.obj.update}
\item{log.sim.change2:}{sim.count}
\item{pair.parameter:}{cell.type.match}
}
\description{
Add.Cell.Cell.Int
}
