% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExprSimulator.R
\name{Add.Distance.Asso.Pattern}
\alias{Add.Distance.Asso.Pattern}
\title{Add.Distance.Asso.Pattern}
\usage{
Add.Distance.Asso.Pattern(
  ppp.obj,
  sim.count,
  r,
  perturbed.cell.type,
  adjacent.cell.type,
  int.dist.threshold = 0.1,
  delta.mean = 1,
  delta.sd = 0.001,
  GeneID = NULL,
  PropOfGenes = NULL,
  seed = NULL
)
}
\arguments{
\item{ppp.obj}{Spatial info for cell type 1 (e.g. neuron)}

\item{sim.count}{Spatial info for cell type 2 (e.g. microglia)}

\item{r}{Expression info for cell type 1 (e.g. neuron)}

\item{perturbed.cell.type}{Expression info for cell type 1 (e.g. microglia)}

\item{adjacent.cell.type}{Int.Cell.Pair.Idx estimated from function \code{Find.Neighbor.Pairs}}

\item{int.dist.threshold}{effect.size}

\item{delta.mean}{effect.size}

\item{delta.sd}{effect.size}

\item{GeneID}{effect.size}

\item{PropOfGenes}{effect.size}

\item{seed}{effect.size}
}
\value{
\item{SignalSummary:}{SignalSummary}
\item{beta.matrix:}{beta.matrix}
}
\description{
This function add cell-cell interactions to a pair of cell types (e.g.
neuron-microglia) for expression in a cell type associated with the
proximity of the other cell type.
One can repeat this function for multiple times to add cell-cell interactions for many cell types.
}
