% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExprSimulato.R
\name{Add.Interactions.to.Pairs}
\alias{Add.Interactions.to.Pairs}
\title{Add.Interactions.to.Pairs}
\usage{
Add.Interactions.to.Pairs(
  pts1,
  pts2,
  sim.count1,
  sim.count2,
  Int.Cell.Pair.Idx,
  Gaussian.effect.size.parameter = c(0, 1),
  mech1.prop = 0.2,
  mech2.prop = 0.1,
  mech3.prop = 0.2,
  seed = NULL
)
}
\arguments{
\item{pts1}{PointLoc}

\item{pts2}{PointLoc}

\item{sim.count1}{PointLoc}

\item{sim.count2}{PointLoc}

\item{Int.Cell.Pair.Idx}{PointLoc}

\item{Gaussian.effect.size.parameter}{PointLoc}
}
\value{
\item{log.sim.change1:}{ppp.obj.update}
\item{log.sim.change2:}{sim.count}
\item{pair.parameter:}{cell.type.match}
}
\description{
Add.Interactions.to.Pairs
}
