% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExprSimulato.R
\name{Add.Polygon.Pattern}
\alias{Add.Polygon.Pattern}
\title{Add.Polygon.Pattern}
\usage{
Add.Polygon.Pattern(
  Step2,
  CellType = "Astrocyte",
  nGrid = 10,
  nPoly = 3,
  PropOfGenes = 0.1,
  Gaussian.effect.size.parameter = c(1, 0.1)
)
}
\description{
Add.Polygon.Pattern
}
