% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExprSimulator.R
\name{Add.Spatial.Expr.Pattern}
\alias{Add.Spatial.Expr.Pattern}
\title{Add.Spatial.Expr.Pattern}
\usage{
Add.Spatial.Expr.Pattern(
  sim.count,
  r,
  CellType,
  GeneID = NULL,
  PropOfGenes = 0.1,
  delta.mean = 1,
  delta.sd = 0.01,
  seed
)
}
\arguments{
\item{sim.count}{Cells as points on a spatial map.}

\item{r}{Region index.}

\item{CellType}{Cell type index.}

\item{GeneID}{Gene(s) index.}

\item{PropOfGenes}{Proportion of genes with this pattern if GeneID is not provided.}

\item{delta.mean}{Mean effect}

\item{delta.sd}{SD of effect.}

\item{seed}{Seed}
}
\value{
Adding spatial differential expressed patterns for one cell type in one region.
}
\description{
This function adds spatial differential expressed patterns for one cell type in one region.
}
