% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointSimulator_NoData.R
\name{cell.loc.fc}
\alias{cell.loc.fc}
\title{cell.loc.fc}
\usage{
cell.loc.fc(
  N,
  win,
  cell.prop,
  cell.inh.attr.input,
  same.dis.cutoff = 0.05,
  even.distribution.coef = 0.1,
  grid.size.small = 19,
  grid.size.large = 45,
  seed
)
}
\arguments{
\item{same.dis.cutoff}{The averaged cell-cell distance is defined as r (e.g. on the unit
square r =1/sqrt(N)). Same cell location is defined as the ratio to r. Default at 0.2.
cell.inh.attr.input=cbind(Cell1=c(1, 3, 4), Cell2=c(2,3,5), Strength= c(2, 0, -2))}

\item{n.vec.raw:}{n.vec.raw from \code{get.n.vec.raw}.}

\item{cell.inh.attr.input:}{cell.inh.attr.input}

\item{same.dis.cutoff:}{same.dis.cutoff =0.05}
}
\description{
This function generates cell locations for each region
}
