% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointSimulator.R
\name{cell.loc.function.fc}
\alias{cell.loc.function.fc}
\title{cell.loc.function.fc}
\usage{
cell.loc.function.fc(
  n.vec,
  density.list,
  window,
  same.loc.force.ratio = 0.05,
  inhibition.attraction.celltype.pair = NULL,
  inhibition.attraction.coef = NULL,
  density.dist.flexibility.coef = c(0, 1, 1, 1, 1),
  even.distribution.coef = 0.1,
  grid.size.small = 20,
  grid.size.large = 45
)
}
\description{
cell.loc.function.fc
}
