% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointSimulator.R
\name{cell.loc.model.fc}
\alias{cell.loc.model.fc}
\title{simulate ST data location based on parametric model}
\usage{
cell.loc.model.fc(n, ppp.obj)
}
\description{
simulate ST data location based on parametric model
}
