% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointSimulator_STData.R
\name{cell.region.loc.model.fc}
\alias{cell.region.loc.model.fc}
\title{simulate ST data location for all regions based on a parametric model}
\usage{
cell.region.loc.model.fc(
  n,
  PointLoc,
  PointAnno,
  PointRegion,
  window_method,
  seed = NULL
)
}
\description{
simulate ST data location for all regions based on a parametric model
}
