% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointSimulator.R
\name{common.density.fc}
\alias{common.density.fc}
\title{several commonly used distribution function}
\usage{
common.density.fc(
  opt.x = c("random", "polynomial", "Gaussian"),
  opt.y = c("random", "polynomial", "Gaussian"),
  poly.coef.x = c(1, 0, 0),
  poly.coef.y = c(1, 0, 0),
  G.mu.x = 0.5,
  G.sd.x = 0.1,
  G.mu.y = 0.5,
  G.sd.y = 0.1
)
}
\description{
several commonly used distribution function
}
