% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointSimulator_NoData.R
\name{connectUp}
\alias{connectUp}
\title{connectUp}
\usage{
connectUp(poly, nPoly, seed = NULL)
}
\arguments{
\item{nPoly:}{nPoly is the No. of regions (e.g. nPoly=3)}

\item{poly:}{poly is a SpatialPolygonsDataFrame (e.g. 20 by 20 square).}
}
\value{
\item{selected.unique:}{A list of the selected polygons for each region.}
}
\description{
This function assigns random connected regions on a square. Used within function \code{RandomRegionWindow}.
}
