% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{contrast.mech1.expr.neighbor}
\alias{contrast.mech1.expr.neighbor}
\title{contrast.mech1.expr.neighbor}
\usage{
contrast.mech1.expr.neighbor(ppp.obj, gene1before, gene1after, neighbor)
}
\description{
contrast.mech1.expr.neighbor
}
