% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{contrast.mech2.expr.neighbor}
\alias{contrast.mech2.expr.neighbor}
\title{contrast.mech2.expr.neighbor}
\usage{
contrast.mech2.expr.neighbor(
  ppp.obj,
  gene1before,
  gene1after,
  gene2before,
  gene2after,
  cell.type = c("ExcitatoryNeuron", "Astrocyte"),
  neighbor
)
}
\description{
contrast.mech2.expr.neighbor
}
