% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{contrast.mech2.expr}
\alias{contrast.mech2.expr}
\title{contrast.mech2.expr}
\usage{
contrast.mech2.expr(
  ppp.obj,
  gene1before,
  gene1after,
  gene2before,
  gene2after,
  cell.type = c("ExcitatoryNeuron", "Astrocyte")
)
}
\description{
contrast.mech2.expr
}
