% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDesign2_fit_revised.R
\name{Est_GeneCopula}
\alias{Est_GeneCopula}
\title{Estimate Gaussian Copula for Gene Expression Matrix}
\usage{
Est_GeneCopula(expr, anno, zp_cutoff = 0.8, min_nonzero_num = 3, ncores = 1)
}
\arguments{
\item{expr}{expr}

\item{anno}{description}

\item{zp_cutoff}{description}

\item{min_nonzero_num}{description}

\item{ncores}{description}
}
\value{
Estimated Gaussian Copula
}
\description{
Estimate Gaussian Copula for Gene Expression Matrix
}
