% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointSimulator_NoData.R
\name{get.n.vec.raw}
\alias{get.n.vec.raw}
\title{get.n.vec.raw}
\usage{
get.n.vec.raw(n, cell.prop, cell.inh.attr.input = NULL, same.dis.cutoff = 0.05)
}
\description{
This function generates cell pools allowing selection due
to cell overlaps, inhibitions and attractions.
}
