% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterDigest.R
\name{loadRData}
\alias{loadRData}
\title{Load RData regardless of the file name}
\usage{
loadRData(fileName)
}
\arguments{
\item{fileName}{File name}
}
\value{
Data
}
\description{
Load RData regardless of the file name
}
