% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExprSimulator.R
\name{MergeRegion}
\alias{MergeRegion}
\title{MergeRegion}
\usage{
MergeRegion(points.list, expr.list)
}
\arguments{
\item{points.list:}{points.list a list of points from multiple regions}

\item{expr.list:}{points.list a list of expressions from multiple regions}
}
\value{
\item{meta:}{meta}
\item{count:}{count}
}
\description{
MergeRegion
}
