% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointSimulator.R
\name{n.cell.type.fc}
\alias{n.cell.type.fc}
\title{Simulate the number of cells per type, with pre-specified cell number K, cell type number K or cell-type proportions.}
\usage{
n.cell.type.fc(n, K = NULL, cell.prop = NULL, lambda = 1, seed = NULL)
}
\description{
Simulate the number of cells per type, with pre-specified cell number K, cell type number K or cell-type proportions.
}
