% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterDigest.R
\name{ParaDigest}
\alias{ParaDigest}
\title{ParaDigest}
\usage{
ParaDigest(input)
}
\arguments{
\item{input}{name for the input parameter file}
}
\value{
Updated parameters
}
\description{
Digest and clean the parameter file.
}
