% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterDigest.R
\name{ParaSimulation}
\alias{ParaSimulation}
\title{(Main Function) Simulate spatially resolved transcriptomics data from a parameter file.}
\usage{
ParaSimulation(input, ModelFitFile = NULL)
}
\arguments{
\item{input}{The path and name of the parameter file.}

\item{ModelFitFile}{Default = NULL, no existing models that fit in the distributions
of input single-cell expression data. Alternatively, if models are provided,
the algorithm will no longer need to fit the input data and be faster.}
}
\value{
Simulated data (e.g. count, spatial feature, expression pattern) will be directly
saved on your computer or cloud based on the path provided by the parameter file.
}
\description{
This function simulate spatially resolved transcriptomics data from a parameter file. The
parameter file can be generated with an user interface on Docker.
}
