% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExprSimulator.R
\name{Pattern.adj.1region}
\alias{Pattern.adj.1region}
\title{Pattern.adj.1region}
\usage{
Pattern.adj.1region(
  sim.count1,
  combined.beta.matrix,
  bond.extreme = T,
  keep.total.count = T,
  integer = T
)
}
\arguments{
\item{sim.count1}{Spatial info for cell type 1 (e.g. neuron)}

\item{combined.beta.matrix}{Spatial info for cell type 2 (e.g. microglia)}

\item{bond.extreme}{Expression info for cell type 1 (e.g. neuron)}

\item{keep.total.count}{Expression info for cell type 1 (e.g. microglia)}

\item{integer}{Int.Cell.Pair.Idx estimated from function \code{Find.Neighbor.Pairs}}
}
\value{
sim.count.update
}
\description{
Adjust the count data for one region based on all input spatial patterns
}
