% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{plot.by.cell}
\alias{plot.by.cell}
\title{plot.by.cell}
\usage{
plot.by.cell(ppp.obj, nrows = 1)
}
\arguments{
\item{ppp.obj}{PointLoc}

\item{nrows}{PointLoc}
}
\value{
Rescaled Point Location
}
\description{
plot.by.cell
}
