% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{plot.gene.pair.fc}
\alias{plot.gene.pair.fc}
\title{plot.gene.pair.fc}
\usage{
plot.gene.pair.fc(
  ppp.obj = pts,
  counts = dat$expr,
  k1 = 1,
  k2 = 2,
  g = dat$gene.pair.idx[[1]][1, ]
)
}
\description{
plot.gene.pair.fc
}
