% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{plot.neighbor.cell.gene.fc}
\alias{plot.neighbor.cell.gene.fc}
\title{plot.neighbor.cell.gene.fc}
\usage{
plot.neighbor.cell.gene.fc(
  ppp.obj = pts,
  expr = dat$expr,
  k = 1,
  g = dat$Cell.gene.idx[[1]][1]
)
}
\description{
plot.neighbor.cell.gene.fc
}
