% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{plot.spatial.gene.fc}
\alias{plot.spatial.gene.fc}
\title{plot.spatial.gene.fc}
\usage{
plot.spatial.gene.fc(
  ppp.obj = pts,
  expr = dat$expr,
  k = 1,
  g = dat$Spatial.gene.idx[[1]]
)
}
\description{
plot.spatial.gene.fc
}
