% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointSimulator_STData.R
\name{simu.window}
\alias{simu.window}
\title{Simulate the window of spatial data}
\usage{
simu.window(
  PointLoc,
  method = c("convex", "convex2", "convex3", "convex5", "rectangle", "network")
)
}
\arguments{
\item{PointLoc}{PointLoc}

\item{method}{method=c("convex", "convex2", "convex3", "convex5", "rectangle", "network")}
}
\description{
This function esimates window using spatial location data of existing cells.
}
