% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExprSimulato.R
\name{Use_scDesign2}
\alias{Use_scDesign2}
\title{Use_scDesign2}
\usage{
Use_scDesign2(
  ppp.obj,
  expr,
  anno = NULL,
  min.cell.type.num = 10,
  input.cell.type = NULL,
  total_count_new = NULL
)
}
\arguments{
\item{ppp.obj}{PointLoc}

\item{expr}{PointLoc}

\item{anno}{PointLoc}

\item{min.cell.type.num}{PointLoc}

\item{input.cell.type}{PointLoc}

\item{total_count_new}{PointLoc
@import scDesign2}
}
\value{
\item{ppp.obj.update:}{ppp.obj.update}
\item{sim.count:}{sim.count}
\item{cell.type.match:}{cell.type.match}
}
\description{
Use_scDesign2
}
