% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExprSimulator.R
\name{Use_scDesign2_1region}
\alias{Use_scDesign2_1region}
\title{Use_scDesign2_1region}
\usage{
Use_scDesign2_1region(
  ppp.obj1,
  Genes,
  model_params,
  depth_simu_ref_ratio = 1,
  cell_type_sel,
  seed,
  sim_method = c("copula", "ind")
)
}
\arguments{
\item{ppp.obj1}{Cells as points on a spatial map.}

\item{Genes}{Gene names of input expression data}

\item{model_params}{Estimated model parameters from reference data.}

\item{depth_simu_ref_ratio}{Relative sequencing depth in comparison of reference data.}

\item{cell_type_sel}{Cell types in reference data.}

\item{seed}{Seed}

\item{sim_method}{Simulate independent genes using 'ind' or correlated genes using 'copula'.}
}
\value{
Simulated count data for a region.
}
\description{
This function uses input parameters to simulate expression for one region.
}
