% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExprSimulator.R
\name{Use_scDesign2_model_params}
\alias{Use_scDesign2_model_params}
\title{Use_scDesign2_model_params}
\usage{
Use_scDesign2_model_params(
  expr,
  feature,
  Copula = NULL,
  sim_method = c("copula", "ind"),
  region_specific_model,
  ncores = 1
)
}
\description{
Use_scDesign2_model_params
}
