% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointSimulator.R
\name{window.fc}
\alias{window.fc}
\title{Generate window for spatial data}
\usage{
\method{window}{fc}(PointLoc = NULL, method = "network")
}
\arguments{
\item{PointLoc}{PointLoc}

\item{method}{method=c("convex", "rectangle", "network")}
}
\value{

}
\description{
This is aa
}
