% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExprSimulator.R
\name{Add.Spatial.Expr.Pattern}
\alias{Add.Spatial.Expr.Pattern}
\title{Add.Spatial.Expr.Pattern}
\usage{
Add.Spatial.Expr.Pattern(
  sim.count,
  r,
  CellType,
  GeneID = NULL,
  PropOfGenes = 0.1,
  delta.mean = 1,
  delta.sd = 0.01,
  seed
)
}
\description{
Add.Spatial.Expr.Pattern for one cell type
}
