% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExprSimulator.R
\name{Use_scDesign2}
\alias{Use_scDesign2}
\title{Use_scDesign2}
\usage{
Use_scDesign2(
  ppp.obj,
  expr,
  anno = NULL,
  Copula = NULL,
  depth_simu_ref_ratio = 1,
  sim_method = c("copula", "ind"),
  seed,
  SaveCopulaName = NULL
)
}
\arguments{
\item{ppp.obj}{PointLoc}

\item{expr}{PointLoc}

\item{anno}{PointLoc}

\item{Copula}{PointLoc}

\item{depth_simu_ref_ratio}{PointLoc}

\item{sim_method}{PointLoc}

\item{seed}{PointLoc}

\item{SaveCopulaName}{PointLoc}
}
\value{
Simulated count data for all regions.
}
\description{
Use_scDesign2
}
