% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExprSimulator.R
\name{Use_scDesign2}
\alias{Use_scDesign2}
\title{Use Revised scDesign2 Functions to Simulate Expression Under the Null.}
\usage{
Use_scDesign2(
  ppp.obj,
  model_params,
  expr,
  feature,
  depth_simu_ref_ratio = 1,
  sim_method = c("copula", "ind"),
  region_specific_model,
  seed
)
}
\arguments{
\item{ppp.obj}{Cells as points on a spatial map for all regions.}

\item{expr}{Gene expression (count) in reference data.}

\item{feature}{Cell features (e.g. cell type, spatial coordinates, regions) of reference data.}

\item{depth_simu_ref_ratio}{Relative sequencing depth in comparison of reference data.}

\item{sim_method}{Simulate independent genes using'ind' or correlated genes using 'copula'.}

\item{seed}{Seed}

\item{Copula}{Default is NULL. If 'copula' method is used, a estimated gene-gene correlation copula should be provided.}
}
\value{
Simulated expression count data for all regions.
}
\description{
This function uses input parameters to simulate expression for all regions.
}
