% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExprSimulator.R
\name{Use_scDesign2_1region}
\alias{Use_scDesign2_1region}
\title{Use_scDesign2_1region}
\usage{
Use_scDesign2_1region(
  ppp.obj1,
  expr,
  Copula = NULL,
  depth_simu_ref_ratio = 1,
  cell_type_sel,
  seed,
  sim_method = c("copula", "ind")
)
}
\arguments{
\item{ppp.obj1}{PointLoc}

\item{expr}{PointLoc}

\item{Copula}{PointLoc}

\item{depth_simu_ref_ratio}{PointLoc}

\item{cell_type_sel}{PointLoc}

\item{seed}{PointLoc}

\item{sim_method}{PointLoc}
}
\value{
Simulated count data for a region.
}
\description{
Use_scDesign2_1region
}
